pyproject.toml
[tool.poetry]
name = "astrology backend"
version = "0.1.4"
description = "Backend астрологического приложения"
authors = ["Adrian Makridenko"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.0.1"
flask-bootstrap = "^3.3.7.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
ymaps = "^1.1"
pyswisseph = "^2.10.3.2"
pytz = "^2023.3"
timezonefinder = "^6.2.0"
tzwhere = "^3.0.3"
flask-restful = "^0.3.10"
httpie = "^3.2.2"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
